+++ src/main/java/eu/printingin3d/physics/HeatingParam.java
@@ -1,11 +1,13 @@
 package eu.printingin3d.physics;
 
+import java.math.BigDecimal;
+
 public class HeatingParam extends DoubleValue {
 	/**
 	 * 
 	 * @param heatingParam in W/K
 	 */
-	public HeatingParam(double heatingParam) {
+	public HeatingParam(BigDecimal heatingParam) {
 		super(heatingParam);
 	}
 }
